Move component types to config folder
