+++ ui/src/implementations/kapture-backend/config/componentTypes.js
